requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
joblib
scikit-learn
xgboost
tensorflow==2.15.0
keras==2.15.0
statsmodels
prophet
pandas
numpy
matplotlib
flask
flask_sqlalchemy
flask_migrate
flask_cors
python-dotenv
flask-bcrypt
flask_wtf
email-validator
pymysql